+++ tests/testthat/test-utils.R
@@ -29,10 +29,8 @@ test_that("respects httr verbose config", {
   expect_equal(httr2_verbosity(), 1)
 })
 
-test_that("can suppress progress bar", {
-  withr::local_options(httr2_progress = FALSE)
-
-  expect_snapshot(sys_sleep(0.1, "for test"))
+test_that("progress bar suppressed in tests", {
+  expect_snapshot(sys_sleep(0.1, "in test"))
 })