+++ NEWS.md
@@ -1,3 +1,5 @@
+# httr2 (development version)
+
 # httr2 1.0.3
 
 * `jwt_encode_hmac()` now calls correct underlying function