+++ NEWS.md
@@ -1,5 +1,6 @@
 # httr2 (development version)
 
+* `req_body_*()` now give informative error if you attempt to change the body type (#451).
 * `resp_body_html()` and `resp_body_xml()` now work when `req_perform()` is given a path (#448).
 * `req_body_file()` now works with files >64kb once more (#524).
 * New `req_perform_connection()` for working with streaming data. Unlike `req_perform_stream()` which uses callbacks, `req_perform_connection()` returns a regular response object with a connection as the body. It's paired with `resp_stream_bytes()`, `resp_stream_lines()`, and `resp_stream_sse()` that allows you to stream chunks as you want them. Unlike `req_perform_stream()` it supports `req_retry()` (with @jcheng5, #519).